generate relevant claims as part of document metadata inference
